+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased] - 2021-xx-xx
+
+### Added
+
+- Support for resend confirmation email by the user during login
+
+#### Changed
+
+- Construct VO membership entitlements if the user is active or in grace period
+
 ## [v2.0.5] - 2021-04-08
 
 ### Added